Extracted DB logic to file, created parsers, added tests for SystemStatCollector
