+++ pg_view/view.py
@@ -3,11 +3,9 @@
 
 from __future__ import absolute_import
 
-import glob
 import logging
 import platform
 import resource
-import socket
 import sys
 import time
 import traceback
@@ -21,21 +19,18 @@ import re
 path = os.path.dirname(os.path.dirname(__file__))
 sys.path.insert(0, path)
 
-if sys.hexversion >= 0x03000000:
-    import configparser as ConfigParser
-else:
-    import ConfigParser
-
-from pg_view.models.base import StatCollector, COLSTATUS, COLALIGN, COLTYPES, COLHEADER, OUTPUT_METHOD, enum, \
-    TICK_LENGTH
+from pg_view.models.base import StatCollector, COLSTATUS, COLALIGN, COLTYPES, COLHEADER, OUTPUT_METHOD, TICK_LENGTH
 import pg_view.models.base
 from pg_view.models.host_stat import HostStatCollector
 from pg_view.models.memory_stat import MemoryStatCollector
 from pg_view.models.partition_stat import PartitionStatCollector, DetachedDiskStatCollector
-from pg_view.models.pg_stat import PgStatCollector, dbversion_as_float
+from pg_view.models.pg_stat import PgStatCollector
 from pg_view.models.system_stat import SystemStatCollector
 from pg_view.validators import get_valid_output_methods, output_method_is_valid
 from pg_view.consumers import DiskCollectorConsumer
+from pg_view.models.db_client import establish_user_defined_connection, make_cluster_desc, build_connection, \
+    read_configuration, detect_db_connection_arguments, get_postmasters_directories
+from pg_view.helpers import process_groups
 
 __appname__ = 'pg_view'
 __version__ = '1.3.0'
@@ -56,10 +51,6 @@ try:
 except ImportError:
     print('Unable to import ncurses, curses output will be unavailable')
     curses_available = False
-
-
-STAT_FIELD = enum(st_pid=0, st_process_name=1, st_state=2, st_ppid=3, st_start_time=21)
-MEM_PAGE_SIZE = resource.getpagesize()
 
 # setup system constants
 output_method = OUTPUT_METHOD.curses
@@ -597,49 +588,6 @@ class CursesOutput(object):
         return layout
 
 
-# some utility functions
-def read_configuration(config_file_name):
-    # read PostgreSQL connection options
-    config_data = {}
-    if not config_file_name:
-        return None
-    config = ConfigParser.ConfigParser()
-    f = config.read(config_file_name)
-    if not f:
-        logger.error('Configuration file {0} is empty or not found'.format(config_file_name))
-        return None
-    # get through all defined databases
-    for section in config.sections():
-        config_data[section] = {}
-        for argname in ('port', 'host', 'user', 'dbname'):
-            try:
-                val = config.get(section, argname)
-            except ConfigParser.NoOptionError:
-                val = None
-            # might happen also if the option is there, but the value is not set
-            if val is not None:
-                config_data[section][argname] = val
-    return config_data
-
-
-def read_postmaster_pid(work_directory, dbname):
-    """ Parses the postgres directory tree and extracts the pid of the postmaster process """
-    fp = None
-    try:
-        fp = open('{0}/postmaster.pid'.format(work_directory))
-        pid = fp.readline().strip()
-    except:
-        # XXX: do not bail out in case we are collecting data for multiple PostgreSQL clusters
-        logger.error('Unable to read postmaster.pid for {name} at {wd}\n HINT: \
-            make sure Postgres is running'.format(name=dbname,
-                     wd=work_directory))
-        return None
-    finally:
-        if fp is not None:
-            fp.close()
-    return pid
-
-
 # execution starts here
 def loop(collectors, consumer, groups, output_method):
     if output_method == OUTPUT_METHOD.curses:
@@ -702,15 +650,15 @@ def do_loop(screen, groups, output_method, collectors, consumer):
     while 1:
         # process input:
         consumer.consume()
-        for st in collectors:
+        for collector in collectors:
             if output_method == OUTPUT_METHOD.curses:
                 if not poll_keys(screen, output):
                     # bail out immediately
                     return
-            st.set_units_display(display_units)
-            st.set_ignore_autohide(not autohide_fields)
-            st.set_notrim(notrim)
-            process_single_collector(st)
+            collector.set_units_display(display_units)
+            collector.set_ignore_autohide(not autohide_fields)
+            collector.set_notrim(notrim)
+            process_single_collector(collector, filter_aux)
             if output_method == OUTPUT_METHOD.curses:
                 if not poll_keys(screen, output):
                     return
@@ -721,8 +669,8 @@ def do_loop(screen, groups, output_method, collectors, consumer):
         # clears the screen, so we need to refresh before display to clear the old data.
         if options.clear_screen and output_method != OUTPUT_METHOD.curses:
             output.refresh()
-        for st in collectors:
-            output.display(st.output(output_method))
+        for collector in collectors:
+            output.display(collector.output(output_method))
         # in the curses case, refresh shows the data queued by display
         if output_method == OUTPUT_METHOD.curses:
             output.refresh()
@@ -730,430 +678,24 @@ def do_loop(screen, groups, output_method, collectors, consumer):
             time.sleep(TICK_LENGTH)
 
 
-def process_single_collector(st):
+def process_single_collector(collector, filter_aux):
     """ perform all heavy-lifting for a single collector, i.e. data collection,
         diff calculation, etc. This is meant to be run in a separate thread.
     """
 
-    if isinstance(st, PgStatCollector):
-        st.set_aux_processes_filter(filter_aux)
-    st.tick()
+    if isinstance(collector, PgStatCollector):
+        collector.set_aux_processes_filter(filter_aux)
+    collector.tick()
     if not freeze:
-        if st.needs_refresh():
-            st.refresh()
-        if st.needs_diffs():
-            st.diff()
+        if collector.needs_refresh():
+            collector.refresh()
+        if collector.needs_diffs():
+            collector.diff()
         else:
             # if the server goes offline, we need to clear diffs here,
             # otherwise rows from the last successful reading will be
             # displayed forever
-            st.clear_diffs()
-
-
-def process_groups(groups):
-    for name in groups:
-        part = groups[name]['partitions']
-        pg = groups[name]['pg']
-        part.ncurses_set_prefix(pg.ncurses_produce_prefix())
-
-
-def get_postmasters_directories():
-    """ detect all postmasters running and get their pids """
-
-    pg_pids = []
-    postmasters = {}
-    pg_proc_stat = {}
-    # get all 'number' directories from /proc/ and sort them
-    for f in glob.glob('/proc/[0-9]*/stat'):
-        # make sure the particular pid is accessible to us
-        if not os.access(f, os.R_OK):
-            continue
-        try:
-            with open(f, 'rU') as fp:
-                stat_fields = fp.read().strip().split()
-        except:
-            logger.error('failed to read {0}'.format(f))
-            continue
-        # read PostgreSQL processes. Avoid zombies
-        if len(stat_fields) < STAT_FIELD.st_start_time + 1 or stat_fields[STAT_FIELD.st_process_name] not in \
-                ('(postgres)', '(postmaster)') or stat_fields[STAT_FIELD.st_state] == 'Z':
-            if stat_fields[STAT_FIELD.st_state] == 'Z':
-                logger.warning('zombie process {0}'.format(f))
-            if len(stat_fields) < STAT_FIELD.st_start_time + 1:
-                logger.error('{0} output is too short'.format(f))
-            continue
-        # convert interesting fields to int
-        for no in STAT_FIELD.st_pid, STAT_FIELD.st_ppid, STAT_FIELD.st_start_time:
-            stat_fields[no] = int(stat_fields[no])
-        pid = stat_fields[STAT_FIELD.st_pid]
-        pg_proc_stat[pid] = stat_fields
-        pg_pids.append(pid)
-
-    # we have a pid -> stat fields map, and an array of all pids.
-    # sort pids array by the start time of the process, so that we
-    # minimize the number of looks into /proc/../cmdline latter
-    # the idea is that processes starting earlier are likely to be
-    # parent ones.
-    pg_pids.sort(key=lambda pid: pg_proc_stat[pid][STAT_FIELD.st_start_time])
-    for pid in pg_pids:
-        st = pg_proc_stat[pid]
-        ppid = st[STAT_FIELD.st_ppid]
-        # if parent is also a postgres process - no way this is a postmaster
-        if ppid in pg_pids:
-            continue
-        link_filename = '/proc/{0}/cwd'.format(pid)
-        # now get its data directory in the /proc/[pid]/cmdline
-        if not os.access(link_filename, os.R_OK):
-            logger.warning('potential postmaster work directory file {0} is not accessible'.format(link_filename))
-            continue
-        # now read the actual directory, check this is accessible to us and belongs to PostgreSQL
-        # additionally, we check that we haven't seen this directory before, in case the check
-        # for a parent pid still produce a postmaster child. Be extra careful to catch all exceptions
-        # at this phase, we don't want one bad postmaster to be the reason of tool's failure for the
-        # other good ones.
-        try:
-            pg_dir = os.readlink(link_filename)
-        except os.error as e:
-            logger.error('unable to readlink {0}: OS reported {1}'.format(link_filename, e))
-            continue
-        if pg_dir in postmasters:
-            continue
-        if not os.access(pg_dir, os.R_OK):
-            logger.warning('unable to access the PostgreSQL candidate directory {0}, have to skip it'.format(pg_dir))
-            continue
-        # if PG_VERSION file is missing, this is not a postgres directory
-        PG_VERSION_FILENAME = '{0}/PG_VERSION'.format(link_filename)
-        if not os.access(PG_VERSION_FILENAME, os.R_OK):
-            logger.warning('PostgreSQL candidate directory {0} is missing PG_VERSION file, have to skip it'.format(
-                           pg_dir))
-            continue
-        try:
-            fp = open(PG_VERSION_FILENAME, 'rU')
-            val = fp.read().strip()
-            if val is not None and len(val) >= 3:
-                version = float(val)
-        except os.error as e:
-            logger.error('unable to read version number from PG_VERSION directory {0}, have to skip it'.format(pg_dir))
-            continue
-        except ValueError:
-            logger.error('PG_VERSION doesn\'t contain a valid version number: {0}'.format(val))
-            continue
-        else:
-            dbname = get_dbname_from_path(pg_dir)
-            postmasters[pg_dir] = [pid, version, dbname]
-    return postmasters
-
-
-def get_dbname_from_path(db_path):
-    """
-    >>> get_dbname_from_path('foo')
-    'foo'
-    >>> get_dbname_from_path('/pgsql_bar/9.4/data')
-    'bar'
-    """
-    m = re.search(r'/pgsql_(.*?)(/\d+.\d+)?/data/?', db_path)
-    if m:
-        dbname = m.group(1)
-    else:
-        dbname = db_path
-    return dbname
-
-
-def build_connection(host, port, user, database):
-    result = {}
-    if host:
-        result['host'] = host
-    if port:
-        result['port'] = port
-    if user:
-        result['user'] = user
-    if database:
-        result['database'] = database
-    return result
-
-
-def fetch_socket_inodes_for_process(pid):
-    """ read /proc/[pid]/fd and get those that correspond to sockets """
-    inodes = []
-    fd_dir = '/proc/{0}/fd'.format(pid)
-    if not os.access(fd_dir, os.R_OK):
-        logger.warning("unable to read {0}".format(fd_dir))
-    else:
-        for link in glob.glob('{0}/*'.format(fd_dir)):
-            if not os.access(link, os.F_OK):
-                logger.warning("unable to access link {0}".format(link))
-                continue
-            try:
-                target = os.readlink(link)
-            except:
-                logger.error('coulnd\'t read link {0}'.format(link))
-            else:
-                # socket:[8430]
-                match = re.search(r'socket:\[(\d+)\]', target)
-                if match:
-                    inodes.append(int(match.group(1)))
-    return inodes
-
-
-def detect_with_postmaster_pid(work_directory, version):
-    # PostgreSQL 9.0 doesn't have enough data
-    result = {}
-    if version is None or version == 9.0:
-        return None
-    PID_FILE = '{0}/postmaster.pid'.format(work_directory)
-
-    # try to access the socket directory
-    if not os.access(work_directory, os.R_OK | os.X_OK):
-        logger.warning('cannot access PostgreSQL cluster directory {0}: permission denied'.format(work_directory))
-        return None
-    try:
-        with open(PID_FILE, 'rU') as fp:
-            lines = fp.readlines()
-    except os.error as e:
-        logger.error('could not read {0}: {1}'.format(PID_FILE, e))
-        return None
-    if len(lines) < 6:
-        logger.error('{0} seems to be truncated, unable to read connection information'.format(PID_FILE))
-        return None
-    port = lines[3].strip()
-    unix_socket_path = lines[4].strip()
-    if unix_socket_path != '':
-        result['unix'] = [(unix_socket_path, port)]
-    tcp_address = lines[5].strip()
-    if tcp_address != '':
-        if tcp_address == '*':
-            tcp_address = '127.0.0.1'
-        result['tcp'] = [(tcp_address, port)]
-    if len(result) == 0:
-        logger.error('could not acquire a socket postmaster at {0} is listening on'.format(work_directory))
-        return None
-    return result
-
-
-def pick_connection_arguments(conn_args):
-    """ go through all decected connections, picking the first one that actually works """
-    result = {}
-    for conn_type in ('unix', 'tcp', 'tcp6'):
-        if len(result) > 0:
-            break
-        for arg in conn_args.get(conn_type, []):
-            if can_connect_with_connection_arguments(*arg):
-                (result['host'], result['port']) = arg
-                break
-    return result
-
-
-def can_connect_with_connection_arguments(host, port):
-    """ check that we can connect given the specified arguments """
-    conn = build_connection(host, port, options.username, options.dbname)
-    try:
-        test_conn = psycopg2.connect(**conn)
-        test_conn.close()
-    except psycopg2.OperationalError:
-        return False
-    return True
-
-
-def detect_with_proc_net(pid):
-    inodes = fetch_socket_inodes_for_process(pid)
-    parser = ProcNetParser()
-    result = parser.match_socket_inodes(inodes)
-    if not result or len(result) == 0:
-        logger.error('could not detect connection string from /proc/net for postgres process {0}'.format(pid))
-        return None
-    return result
-
-
-def detect_db_connection_arguments(work_directory, pid, version):
-    """
-        Try to detect database connection arguments from the postmaster.pid
-        We do this by first extracting useful information from postmaster.pid,
-        next reading the postgresql.conf if necessary and, at last,
-    """
-    conn_args = detect_with_proc_net(pid)
-    if not conn_args:
-        # if we failed to detect the arguments via the /proc/net/ readings,
-        # perhaps we'll get better luck with just peeking into postmaster.pid.
-        conn_args = detect_with_postmaster_pid(work_directory, version)
-        if not conn_args:
-            logger.error('unable to detect connection parameters for the PostgreSQL cluster at {0}'.format(
-                         work_directory))
-            return None
-    # try all acquired connection arguments, starting from unix, then tcp, then tcp over ipv6
-    result = pick_connection_arguments(conn_args)
-    if len(result) == 0:
-        logger.error('unable to connect to PostgreSQL cluster ' +
-                     'at {0} using any of the detected connection options: {1}'.format(work_directory, conn_args))
-        return None
-    return result
-
-
-def establish_user_defined_connection(instance, conn, clusters):
-    """ connect the database and get all necessary options like pid and work_directory
-        we use port, host and socket_directory, prefering socket over TCP connections
-    """
-    # establish a new connection
-    try:
-        pgcon = psycopg2.connect(**conn)
-    except Exception as e:
-        logger.error('failed to establish connection to {0} via {1}'.format(instance, conn))
-        logger.error('PostgreSQL exception: {0}'.format(e))
-        return None
-    # get the database version from the pgcon properties
-    dbver = dbversion_as_float(pgcon)
-    cur = pgcon.cursor()
-    cur.execute('show data_directory')
-    work_directory = cur.fetchone()[0]
-    cur.close()
-    pgcon.commit()
-    # now, when we have the work directory, acquire the pid of the postmaster.
-    pid = read_postmaster_pid(work_directory, instance)
-    if pid is None:
-        logger.error('failed to read pid of the postmaster on {0}'.format(conn))
-        return None
-    # check that we don't have the same pid already in the accumulated results.
-    # for instance, a user may specify 2 different set of connection options for
-    # the same database (one for the unix_socket_directory and another for the host)
-    pids = [opt['pid'] for opt in clusters if 'pid' in opt]
-    if pid in pids:
-        duplicate_instance = [opt['name'] for opt in clusters if 'pid' in opt and opt.get('pid', 0) == pid][0]
-        logger.error('duplicate connection options detected ' +
-                     'for databases {0} and {1}, same pid {2}, skipping {0}'.format(instance, duplicate_instance, pid))
-        pgcon.close()
-        return True
-    # now we have all components to create a cluster descriptor
-    desc = make_cluster_desc(
-        name=instance, version=dbver, workdir=work_directory, pid=pid, pgcon=pgcon, conn=conn)
-    clusters.append(desc)
-    return True
-
-
-def make_cluster_desc(name, version, workdir, pid, pgcon, conn):
-    """Create cluster descriptor, complete with the reconnect function."""
-
-    def reconnect():
-        pgcon = psycopg2.connect(**conn)
-        pid = read_postmaster_pid(workdir, name)
-        return pgcon, pid
-
-    return {
-        'name': name,
-        'ver': version,
-        'wd': workdir,
-        'pid': pid,
-        'pgcon': pgcon,
-        'reconnect': reconnect
-    }
-
-
-class ProcNetParser():
-    """ Parse /proc/net/{tcp,tcp6,unix} and return the list of address:port
-        pairs given the set of socket descriptors belonging to the object.
-        The result is grouped by the socket type in a dictionary.
-    """
-    NET_UNIX_FILENAME = '/proc/net/unix'
-    NET_TCP_FILENAME = '/proc/net/tcp'
-    NET_TCP6_FILENAME = '/proc/net/tcp6'
-
-    def __init__(self):
-        self.reinit()
-
-    def reinit(self):
-        self.sockets = {}
-        self.unix_socket_header_len = 0
-        # initialize the sockets hash with the contents of unix
-        # and tcp sockets. tcp IPv6 is also read if it's present
-        for fname in (self.NET_UNIX_FILENAME, self.NET_TCP_FILENAME):
-            self.read_socket_file(fname)
-        if os.access(self.NET_TCP6_FILENAME, os.R_OK):
-            self.read_socket_file(self.NET_TCP6_FILENAME)
-
-    @staticmethod
-    def _hex_to_int_str(val):
-        return str(int(val, 16))
-
-    @staticmethod
-    def _hex_to_ip(val):
-        newval = format(socket.ntohl(int(val, 16)), '08X')
-        return '.'.join([str(int(newval[i: i + 2], 16)) for i in range(0, 8, 2)])
-
-    @staticmethod
-    def _hex_to_ipv6(val):
-        newval_list = [format(socket.ntohl(int(val[x: x + 8], 16)), '08X') for x in range(0, 32, 8)]
-        return ':'.join([':'.join((x[:4], x[4:])) for x in newval_list])
-
-    def match_socket_inodes(self, inodes):
-        """ return the dictionary with socket types as strings,
-            containing addresses (or unix path names) and port
-        """
-        result = {}
-        for inode in inodes:
-            if inode in self.sockets:
-                addr_tuple = self.parse_single_line(inode)
-                if addr_tuple is None:
-                    continue
-                socket_type = addr_tuple[0]
-                if socket_type in result:
-                    result[socket_type].append(addr_tuple[1:])
-                else:
-                    result[socket_type] = [addr_tuple[1:]]
-        return result
-
-    def read_socket_file(self, filename):
-        """ read file content, produce a dict of socket inode -> line """
-        socket_type = filename.split('/')[-1]
-        try:
-            with open(filename) as fp:
-                data = fp.readlines()
-        except os.error as e:
-            logger.error('unable to read from {0}: OS reported {1}'.format(filename, e))
-        # remove the header
-        header = (data.pop(0)).split()
-        if socket_type == 'unix':
-            self.unix_socket_header_len = len(header)
-        indexes = [i for i, name in enumerate(header) if name.lower() == 'inode']
-        if len(indexes) != 1:
-            logger.error('attribute \'inode\' in the header of {0} is not unique or missing: {1}'.format(
-                         filename, header))
-        else:
-            inode_idx = indexes[0]
-            if socket_type != 'unix':
-                # for a tcp socket, 2 pairs of fields (tx_queue:rx_queue and tr:tm->when
-                # are separated by colons and not spaces)
-                inode_idx -= 2
-            for line in data:
-                fields = line.split()
-                inode = int(fields[inode_idx])
-                self.sockets[inode] = [socket_type, line]
-
-    def parse_single_line(self, inode):
-        """ apply socket-specific parsing rules """
-        result = None
-        (socket_type, line) = self.sockets[inode]
-        if socket_type == 'unix':
-            # we are interested in everything in the last field
-            # note that it may contain spaces or other separator characters
-            fields = line.split(None, self.unix_socket_header_len - 1)
-            socket_path = fields[-1]
-            # check that it looks like a PostgreSQL socket
-            match = re.search(r'(.*?)/\.s\.PGSQL\.(\d+)$', socket_path)
-            if match:
-                # path - port
-                result = (socket_type,) + match.groups(1)
-            else:
-                logger.warning('unix socket name is not recognized as belonging to PostgreSQL: {0}'.format(socket_path))
-        else:
-            address_port = line.split()[1]
-            (address_hex, port_hex) = address_port.split(':')
-            port = self._hex_to_int_str(port_hex)
-            if socket_type == 'tcp6':
-                address = self._hex_to_ipv6(address_hex)
-            elif socket_type == 'tcp':
-                address = self._hex_to_ip(address_hex)
-            else:
-                logger.error('unrecognized socket type: {0}'.format(socket_type))
-            result = (socket_type, address, port)
-        return result
+            collector.clear_diffs()
 
 
 def main():
@@ -1216,7 +758,7 @@ def main():
                 if user_dbver is not None and dbver != user_dbver:
                     continue
             try:
-                conndata = detect_db_connection_arguments(result_work_dir, ppid, dbver)
+                conndata = detect_db_connection_arguments(result_work_dir, ppid, dbver, options.username, options.dbname)
                 if conndata is None:
                     continue
                 host = conndata['host']
@@ -1227,8 +769,10 @@ def main():
                 pg_view.models.base.logger.error('PostgreSQL exception {0}'.format(e))
                 pgcon = None
             if pgcon:
-                desc = make_cluster_desc(name=dbname, version=dbver, workdir=result_work_dir,
-                                         pid=ppid, pgcon=pgcon, conn=conn)
+                desc = make_cluster_desc(
+                    name=dbname, version=dbver, workdir=result_work_dir,
+                    pid=ppid, pgcon=pgcon, conn=conn
+                )
                 clusters.append(desc)
 
     collectors = []
